requirements.txt
langchain
langgraph
openai
pandas
pandas_ta
vectorbt
kiteconnect
duckdb
python-dotenv
pyarrow
duckdb
fastparquet